+++ CHANGELOG.md
@@ -9,10 +9,10 @@ and this project adheres to
 
 # Unreleased
 
-# 13.3.0 - 2024-07-16
-
 - Added a `toMatchDataModelSchema` Jest matcher for validating entities against
   the live production data model schema.
+- Deprecated the `toMatchGraphObjectSchema` method in favor of
+  `toMatchDataModelSchema`.
 
 # 13.2.0 - 2024-07-11